Standardize with cdk-bare-bones
https://github.com/symphoniacloud/cdk-bare-bones provides a "bare bones" CDK template. This change incorporates all of the various decisions I made in that
project, e.g. run scripts in package.json.
