+++ .gitignore
@@ -1,2 +1,2 @@
-cdk.out
 node_modules
+build
