+++ cdk.json
@@ -1,3 +1,3 @@
 {
-  "app": "npx ts-node --prefer-ts-exts src/cdk/cdkApp.ts"
+  "app": "npx ts-node --prefer-ts-exts src/cdk/App.ts"
 }
